Refactor style layers to separate file
